Changed Profile Menu to own component
