+++ components/index.ts
@@ -9,3 +9,4 @@ export { default as SignIn } from './SignIn'
 export { default as SignOut } from './SignOut'
 export { default as Meta } from './Meta'
 export { default as Share } from './Share'
+export { default as ProfileMenu } from './Sidebar/ProfileMenu'
